cloud_function/requirements.txt
# Function dependencies, for example:
# package>=version

pandas
fsspec
gcsfs
google-cloud-storage